requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
addict=2.4.0=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.8.30=h56e8100_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.3.2=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.3.0=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.6.7=py312hd77b12b_0
decorator=5.1.1=pyhd8ed1ab_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
executing=2.1.0=pyhd8ed1ab_0
expat=2.6.2=hd77b12b_0
filelock=3.15.4=pypi_0
fonttools=4.53.1=pypi_0
fsspec=2024.6.1=pypi_0
huggingface-hub=0.24.6=pypi_0
idna=3.8=pypi_0
importlib-metadata=8.4.0=pyha770c72_0
importlib_metadata=8.4.0=hd8ed1ab_0
ipykernel=6.29.5=pyh4bbf305_0
ipython=8.27.0=pyh7428d3b_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=pypi_0
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.7.2=py312haa95532_0
kiwisolver=1.4.7=pypi_0
libffi=3.4.4=hd77b12b_1
libsodium=1.0.18=h8d14728_1
markupsafe=2.1.5=pypi_0
matplotlib=3.9.2=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mpmath=1.3.0=pypi_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.3=pypi_0
numpy=1.26.4=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.0.14=h827c3e9_0
packaging=24.1=pyhd8ed1ab_0
parso=0.8.4=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.4.0=pypi_0
pip=24.2=py312haa95532_0
platformdirs=4.2.2=pyhd8ed1ab_0
prompt-toolkit=3.0.47=pyha770c72_0
psutil=5.9.0=py312h2bbff1b_0
pure_eval=0.2.3=pyhd8ed1ab_0
pycocotools=2.0.8=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=3.1.4=pypi_0
python=3.12.4=h14ffc60_1
python-dateutil=2.9.0.post0=pypi_0
pywin32=305=py312h2bbff1b_0
pyyaml=6.0.2=pypi_0
pyzmq=25.1.2=py312hd77b12b_0
regex=2024.7.24=pypi_0
requests=2.32.3=pypi_0
safetensors=0.4.4=pypi_0
setuptools=72.1.0=py312haa95532_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.45.3=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
supervision=0.6.0=pypi_0
sympy=1.13.2=pypi_0
timm=1.0.9=pypi_0
tk=8.6.14=h0416ee5_0
tokenizers=0.19.1=pypi_0
tomli=2.0.1=pypi_0
torch=2.4.0+cu124=pypi_0
torchaudio=2.4.0+cu124=pypi_0
torchvision=0.19.0=pypi_0
tornado=6.4.1=py312h827c3e9_0
tqdm=4.66.5=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
transformers=4.44.2=pypi_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.2=pypi_0
vc=14.40=h2eaa2aa_0
vs2015_runtime=14.40.33807=h98bb1dd_0
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.43.0=py312haa95532_0
xz=5.4.6=h8cc25b3_1
yapf=0.40.2=pypi_0
zeromq=4.3.5=hd77b12b_0
zipp=3.20.1=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_1
